fix : primary-mypage info item 수정
