+++ src/screens/primary/primary-my-page/module/primary-mypage-content/components/index.ts
@@ -1,0 +1,1 @@
+export * from './info-item';
